pyproject.toml
[project]
name = "scheduling-agent"
version = "0.1.0"
description = "AI Email & Scheduling Agent with MCP"
requires-python = ">=3.11"

dependencies = [
    "mcp>=1.0.0",
    "fastmcp>=0.1.0",
    "langgraph>=0.2.0",
    "langchain-openai>=0.2.0",
    "langchain-core>=0.3.0",
    "langchain-mcp-adapters>=0.1.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
]
